% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_ssd7.R
\name{build_model}
\alias{build_model}
\title{Build the SSD7 Model}
\usage{
build_model(image_size, n_classes, min_scale = 0.1, max_scale = 0.9,
  scales = None, aspect_ratios_global = c(0.5, 1, 2),
  aspect_ratios_per_layer = NULL, two_boxes_for_ar1 = TRUE,
  limit_boxes = TRUE, variances = c(1, 1, 1, 1), coords = "centroids",
  normalize_coords = FALSE)
}
\arguments{
\item{a}{a}
}
\description{
Description
}
\examples{
build_model()
}
